Allow non-scalar value on serialize field
If a field has a serialization like "json", "serialize" set, it may have non-scalar values. At least I store serialized values in fields of 'type' => 'text'.
This PR adds an additional condition in normalize()'s check for scalar values. If $this->serialize is set, value does not need to be scalar as its made scalar by serialization on save.